[UPDATE]修改员工信息修改接口

+++ internal/dao/employee_dao.go
@@ -32,7 +32,7 @@ func (m *EmployeeDao) EmployeeSave(employee *pojo.Employee) error {
 	return m.Orm.Create(employee).Error
 }
 
-func (m *EmployeeDao) FindEmployeeById(id int) (pojo.Employee, error) {
+func (m *EmployeeDao) FindEmployeeById(id int64) (pojo.Employee, error) {
 	var employee pojo.Employee
 	err := m.Orm.Where("id = ?", id).First(&employee).Error
 	return employee, err
